refactor(package): 유틸 패키지 제거
