+++ src/main/java/wordle/model/WordPoolGenerator.java
@@ -1,4 +1,4 @@
-package wordle.util;
+package wordle.model;
 
 import java.io.IOException;
 import java.nio.file.Files;
