+++ src/test/java/wordle/model/WordValidatorTest.java
@@ -1,4 +1,4 @@
-package wordle.util;
+package wordle.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import wordle.model.WordPool;
+import wordle.model.WordValidator;
 
 class WordValidatorTest {
